Made a new, more stable basis for our todo testing.
When looking our the tests within the "todoTesting!" branch, I noticed that there existed some errors in compiling some of the tests and getting the status filter to work properly in todos. Therefore, I reverted to a stable build that still held complete functionality for the status keyword and modified it to create a more stable base for implementing todo tests.